requirements.txt
langchain
langchain_core
langchain_groq
python-dotenv
langchain-community
lxml
sentence_transformers
langchain_huggingface
faiss-cpu
pypdf
streamlit
bs4